fix(daemon): isolate the repair-transaction fixtures from the shared repair-loop fixture
Adversarial review caught that widening session-replay-repair.fixtures.ts
with the transaction-specific setup()/handleCloseCommand()/
makeCompleteRepairSession() helpers pulled application-lifecycle-runtime-
harness.ts's node:timers/promises mock into three sibling test files
(session-replay-repair-loop, -empty-tail, -acceptance) that only needed the
fixture's existing makeRecordingReplayInvoke and never declared that mock
themselves. Move those helpers into a new session-replay-repair-
transaction.fixtures.ts sibling instead, so session-replay-repair.fixtures.ts
goes back to its original content and only the three
session-replay-repair-transaction*.test.ts files pull in the close-lifecycle
harness.

Also tightens two stale comments the split left behind: find-handler-
fixture.ts no longer motivates itself by find.test.ts's now-resolved size
tripwire, and find-production-route.test.ts no longer calls the stub-based
tests "above" now that they live in a different file.
